backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.22
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3

